backend_Project/requirements.txt
Django>=4.2,<5
djangorestframework>=3.14,<4
django-cors-headers>=4,<5
psycopg[binary]>=3.1,<4
stripe>=8,<9